// modules/apps/export-import/export-import-rest-client-js/src/models/Creator.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

			import {UserGroupBrief} from './UserGroupBrief';

/**
 * @author Petteri Karttunen
 * @generated
 */

	/**
	* Represents the user account of the content's creator/author. Properties follow the [creator](https://schema.org/creator) specification.
	*/
	export class Creator {
			"additionalName"?: string;
			"contentType"?: string;
			"externalReferenceCode"?: string;
			"familyName"?: string;
			"givenName"?: string;
			"id"?: number;
			"image"?: string;
			"name"?: string;
			"profileURL"?: string;
			"userGroupBriefs"?: Array<UserGroupBrief>;

		static "discriminator": string | undefined = undefined;

	static "attributeTypeMap": Array<{
		baseName: string;
		name: string;
		type: string;
	}> = [
		{
			baseName: "additionalName",
			name: "additionalName",
			type: "string",
		},
		{
			baseName: "contentType",
			name: "contentType",
			type: "string",
		},
		{
			baseName: "externalReferenceCode",
			name: "externalReferenceCode",
			type: "string",
		},
		{
			baseName: "familyName",
			name: "familyName",
			type: "string",
		},
		{
			baseName: "givenName",
			name: "givenName",
			type: "string",
		},
		{
			baseName: "id",
			name: "id",
			type: "number",
		},
		{
			baseName: "image",
			name: "image",
			type: "string",
		},
		{
			baseName: "name",
			name: "name",
			type: "string",
		},
		{
			baseName: "profileURL",
			name: "profileURL",
			type: "string",
		},
		{
			baseName: "userGroupBriefs",
			name: "userGroupBriefs",
			type: "Array<UserGroupBrief>",
		},
		];

		static getAttributeTypeMap() {
				return Creator.attributeTypeMap;
		}
	}